+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 
 * Rename ``file`` argument to ``vcf`` for :command:`vcf-slice` command.
 * Add new command :command:`vcf-index`.
+* Add new method :meth:`pyvcf.has_chr_prefix`.
+* Add new command :meth:`common.update_chr_prefix`.
 
 0.26.0 (2021-10-24)
 -------------------